[MOB-3807] Update the views to be managed by a single generic "error view"
